+++ src/services/metrics/mod.rs
@@ -418,7 +418,7 @@ mod tests {
             .await;
         });
 
-        tokio::time::sleep(Duration::from_secs(NUM_OF_SCRAPES as u64)).await;
+        tokio::time::sleep(Duration::from_secs(u64::try_from(NUM_OF_SCRAPES).unwrap())).await;
         is_shutdown.store(true, Ordering::Release);
         data_receiver.close();
 
@@ -484,7 +484,7 @@ mod tests {
             .await;
         });
 
-        tokio::time::sleep(Duration::from_secs(NUM_OF_SCRAPES as u64)).await;
+        tokio::time::sleep(Duration::from_secs(u64::try_from(NUM_OF_SCRAPES).unwrap())).await;
         is_shutdown.store(true, Ordering::Release);
         data_receiver.close();