+++ src/services/system/mod.rs
@@ -336,7 +336,10 @@ mod tests {
             .await;
         });
 
-        tokio::time::sleep(Duration::from_secs(2 * NUM_OF_SCRAPES as u64)).await;
+        tokio::time::sleep(Duration::from_secs(
+            2 * u64::try_from(NUM_OF_SCRAPES).unwrap(),
+        ))
+        .await;
         is_shutdown.store(true, Ordering::Release);
         data_receiver.close();